Updated Response File Processing
Added in-program display of CSV file after conversion, and a button to
clear the text box on click.

Text box data has no effect on what will be printed to the record file.
VE record will ALWAYS be printed first, and VA records will ALWAYS be
printed in [0..n] order.